Add ToObject to string extensions and its UT. Refactor the test classes to their own folder.
