+++ src/Avalon.Base.Extension/Avalon.Base.Extension.UT/Objects/User.cs
@@ -1,0 +1,13 @@
+﻿namespace Avalon.Base.Extension.UT.Objects;
+
+/// <summary>
+/// Demo class for testing purposes.
+/// </summary>
+internal class User
+{
+    public string Id { get; set; }
+    public string Name { get; set; }
+    public string LastName { get; set; }
+    public string Description { get; set; }
+    public string Password { get; set; }
+}
