+++ src/Avalon.Base.Extension/Avalon.Base.Extension.UT/Objects/UserDTO.cs
@@ -1,0 +1,12 @@
+﻿namespace Avalon.Base.Extension.UT.Objects;
+
+/// <summary>
+/// Demo DTO class for testing purposes.
+/// </summary>
+internal class UserDTO
+{
+    public string Id { get; set; }
+    public string Name { get; set; }
+    public string LastName { get; set; }
+    public string Description { get; set; }
+}
