+++ src/Avalon.Base.Extension/Avalon.Base.Extension.UT/Types/ObjectExtensions/ObjectExtensionsTest.cs
@@ -1,4 +1,5 @@
 ﻿using Avalon.Base.Extension.Types;
+using Avalon.Base.Extension.UT.Objects;
 
 namespace Avalon.Base.Extension.UT.Types.ObjectExtensions;
 
@@ -31,32 +32,8 @@ public class ObjectExtensionsTest
             Description = "User tester",
             Password = "Password1"
         };
-
         UserDTO dto = user.ToDTO<UserDTO>();
 
         Assert.IsTrue(dto.Name.IsNotNullOrEmpty());
     }
-}
-
-/// <summary>
-/// Demo class for testing purposes.
-/// </summary>
-internal class User
-{
-    public string Id { get; set; }
-    public string Name { get; set; }
-    public string LastName { get; set; }
-    public string Description { get; set; }
-    public string Password { get; set; }
-}
-
-/// <summary>
-/// Demo DTO class for testing purposes.
-/// </summary>
-internal class UserDTO
-{
-    public string Id { get; set; }
-    public string Name { get; set; }
-    public string LastName { get; set; }
-    public string Description { get; set; }
 }
